Change model_detector.py to model_manager.py, implement logic to set splash screen
