+++ src/asmcnc/comms/localization.py
@@ -7,7 +7,7 @@ from datetime import datetime
 from kivy.core.text import LabelBase
 from kivy.lang import Builder
 
-from asmcnc.comms import model_detector
+from asmcnc.comms import model_manager
 
 asmcnc_path = os.path.dirname(os.path.dirname(__file__))
 fonts_path = os.path.join(asmcnc_path, "keyboard", "fonts")
